// frontend/src/app/models/connection.ts
export default class connection {
  _id: string;
  user_id: string;
  address: string;
  district: string;
  connection_type: string;
  status: string;
  address_proof: string;
  address_proof_type: string;
  user_id_proof: string;
  user_id_proof_type: string;
  request_date: String;
  connection_date: String;
  current_reading: number;
  balance: number;
}